ConvolutionalNNetwork/requirements.txt
aiohttp==2.3.6
async-timeout==2.0.0
bleach==1.5.0
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
html5lib==0.9999999
imageio==2.2.0
Keras==2.1.2
Markdown==2.6.10
matplotlib==2.1.1
multidict==3.3.2
networkx==2.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
protobuf==3.5.0.post1
pydot==1.2.3
pyparsing==2.2.0
python-dateutil==2.6.1
python-engineio==2.0.1
python-socketio==1.8.4
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scipy==1.0.0
six==1.11.0
tensorflow==1.4.0
tensorflow-tensorboard==0.4.0rc3
Werkzeug==0.13
yarl==0.16.0
